Updated iText to 5.2.1
git-svn-id: https://fb2pdf.googlecode.com/svn/trunk/src/java@996 f05116cd-5028-0410-9863-11db5debc633

+++ src/com/itextpdf/text/pdf/EnumerateTTC.java
@@ -1,8 +1,8 @@
 /*
- * $Id: EnumerateTTC.java 4784 2011-03-15 08:33:00Z blowagie $
+ * $Id: EnumerateTTC.java 5075 2012-02-27 16:36:18Z blowagie $
  *
  * This file is part of the iText (R) project.
- * Copyright (c) 1998-2011 1T3XT BVBA
+ * Copyright (c) 1998-2012 1T3XT BVBA
  * Authors: Bruno Lowagie, Paulo Soares, et al.
  *
  * This program is free software; you can redistribute it and/or modify
@@ -78,7 +78,7 @@ class EnumerateTTC extends TrueTypeFont{
             rf.skipBytes(4);
             int dirCount = rf.readInt();
             names = new String[dirCount];
-            int dirPos = rf.getFilePointer();
+            int dirPos = (int)rf.getFilePointer();
             for (int dirIdx = 0; dirIdx < dirCount; ++dirIdx) {
                 tables.clear();
                 rf.seek(dirPos);
