+++ src/com/itextpdf/text/pdf/OutputStreamCounter.java
@@ -1,8 +1,8 @@
 /*
- * $Id: OutputStreamCounter.java 4784 2011-03-15 08:33:00Z blowagie $
+ * $Id: OutputStreamCounter.java 5075 2012-02-27 16:36:18Z blowagie $
  *
  * This file is part of the iText (R) project.
- * Copyright (c) 1998-2011 1T3XT BVBA
+ * Copyright (c) 1998-2012 1T3XT BVBA
  * Authors: Bruno Lowagie, Paulo Soares, et al.
  *
  * This program is free software; you can redistribute it and/or modify
@@ -52,7 +52,7 @@ import java.io.OutputStream;
 public class OutputStreamCounter extends OutputStream {
     
     protected OutputStream out;
-    protected int counter = 0;
+    protected long counter = 0;
     
     /** Creates a new instance of OutputStreamCounter */
     public OutputStreamCounter(OutputStream out) {
@@ -157,7 +157,7 @@ public class OutputStreamCounter extends OutputStream {
         out.write(b, off, len);
     }
     
-    public int getCounter() {
+    public long getCounter() {
         return counter;
     }
