+++ src/com/itextpdf/text/pdf/PdfStream.java
@@ -1,8 +1,8 @@
 /*
- * $Id: PdfStream.java 4784 2011-03-15 08:33:00Z blowagie $
+ * $Id: PdfStream.java 5075 2012-02-27 16:36:18Z blowagie $
  *
  * This file is part of the iText (R) project.
- * Copyright (c) 1998-2011 1T3XT BVBA
+ * Copyright (c) 1998-2012 1T3XT BVBA
  * Authors: Bruno Lowagie, Paulo Soares, et al.
  *
  * This program is free software; you can redistribute it and/or modify
@@ -335,7 +335,7 @@ public class PdfStream extends PdfDictionary {
             }
             if (ose != null)
                 ose.finish();
-            inputStreamLength = osc.getCounter();
+            inputStreamLength = (int)osc.getCounter();
         }
         else {
             if (crypto != null && !crypto.isEmbeddedFilesOnly()) {
